+++ postmarker/models/emails.py
@@ -50,10 +50,17 @@ def prepare_attachments(attachment):
         if len(attachment) == 4:
             result['ContentID'] = attachment[3]
     elif isinstance(attachment, MIMEBase):
+        payload = attachment.get_payload()
+        content_type = attachment.get_content_type()
+        # Special case for message/rfc822
+        # Even if RFC implies such attachments being not base64-encoded,
+        # Postmark requires all attachments to be encoded in this way
+        if content_type == 'message/rfc822' and not isinstance(payload, str):
+            payload = b64encode(payload[0].get_payload(decode=True)).decode()
         result = {
-            'Name': attachment.get_filename(),
-            'Content': attachment.get_payload(),
-            'ContentType': attachment.get_content_type(),
+            'Name': attachment.get_filename() or 'attachment.txt',
+            'Content': payload,
+            'ContentType': content_type,
         }
         content_id = attachment.get('Content-ID')
         if content_id:
@@ -91,6 +98,11 @@ def deconstruct_multipart_recursive(seen, text, html, attachments, message):
         elif content_type == 'text/html' and not html:
             html.append(message.get_payload(decode=True).decode('utf8'))
         else:
+            # Ignore underlying messages inside `message/rfc822` payload, because the message itself will be passed
+            # as an attachment
+            if content_type == 'message/rfc822' and sys.version_info[:2] not in ((2, 6), (3, 2)):
+                for part in message.get_payload():
+                    seen.add(part)
             attachments.append(message)